Fix handling of partially replicated snapshots
Cover cases where a new snapshot has been taken since the last failure,
in which case most_recent_snapshot would not be the relevant one to
look at. incremental_base on the other hand, is always relevant.
Also in this case we need to run the check even if though 'snapshots'
may not be equal to 'None' (because of the newly taken snapshot)

Also add "-r" to the suggested command to run. Indeed, some children
may have been replicated before the failure, and must be deleted too