NAS-113577 / 22.02-RC.2 / Fix handling of partially replicated snapshots (#192)
* Fix handling of partially replicated snapshots

Cover cases where a new snapshot has been taken since the last failure,
in which case most_recent_snapshot would not be the relevant one to
look at. incremental_base on the other hand, is always relevant.
Also in this case we need to run the check even if though 'snapshots'
may not be equal to 'None' (because of the newly taken snapshot)

Also add "-r" to the suggested command to run. Indeed, some children
may have been replicated before the failure, and must be deleted too

* Fix typo

* Update error message in test

* Add test case covering when a new snapshot has been taken

* Remove obsolete optimization

We always need to query createtxg in order to get an accurate value for
incremental_base, which is required to check for inconsistent replication
states due to a past replication that would have failed halfway.

(cherry picked from commit 5e768d7011c93cb771622253e603a1493c4eb098)