Fix periodic snapshot task snapshot lifetime
It it is `06:00` and snaspshot lifetime is `1 hour` then we should remove all snapshots
that were created at 05:00 or earlier (or retain all that were `created at > 05:00`).
So, the snapshot created at 05:00 should be removed at 06:00.

Previously we were retaining all snapshots that were `created at >= 05:00` (thus, removing
all that were `created at < 05:00`). So the snapshot created at 05:00 was not removed at 06:00.